[BE2-031] Refactored Reset Token files and usage
